+++ packages/eds-core-react/.storybook/components/index.ts
@@ -1,3 +1,3 @@
 export { PropsTable } from './PropsTable'
-export { Story } from './../Story'
-export { Stack } from './../Stack'
+export { Story } from './Story'
+export { Stack } from './Stack'
